Update scripts to use int instead of byte   - Avoid data unalignment (https://web.archive.org/web/20080607055623/http://www.ibm.com/developerworks/library/pa-dalign/).   - Avoid having to cast operations (C# byte + byte = int).   - Be standard.   - Avoid premature optimization.

+++ Assets/Scripts/Board/BoardCoordinate.cs
@@ -1,9 +1,9 @@
 public readonly struct BoardCoordinate
 {
-    public readonly byte row;
-    public readonly byte column;
+    public readonly int row;
+    public readonly int column;
 
-    public BoardCoordinate(in byte row, in byte column)
+    public BoardCoordinate(in int row, in int column)
     {
         this.row = row;
         this.column = column;
